# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.FormStatus:
    enum:
    - DRAFT
    - ACTIVE
    - CLOSED
    type: string
    x-enum-varnames:
    - StatusDraft
    - StatusActive
    - StatusClosed
  domain.FormType:
    enum:
    - SURVEY
    - EXAM
    type: string
    x-enum-varnames:
    - TypeSurvey
    - TypeExam
...
  domain.QuestionType:
    enum:
    - SHORT_TEXT
    - LONG_TEXT
    - MULTIPLE_CHOICE
    - CHECKBOXES
    - DROPDOWN
    - RATING
    - YES_NO
    - MATH
    - CODE
    - IMAGE
    - MATCHING
    type: string
    x-enum-varnames:
    - TypeShortText
    - TypeLongText
    - TypeMultipleChoice
    - TypeCheckboxes
    - TypeDropdown
    - TypeRating
    - TypeYesNo
    - TypeMath
    - TypeCode
    - TypeImage
    - TypeMatching
  domain.UserRole:
    enum:
    - superadmin
    - admin
    - user
    type: string
    x-enum-varnames:
    - RoleSuperAdmin
    - RoleAdmin
    - RoleUser
...
  dto.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.MatchPairItem:
    properties:
      match_key:
        type: string
      match_target_text:
        type: string
    type: object
...
  dto.PublicFormDTO:
    properties:
      category:
        type: string
      custom_url:
        type: string
      description:
        type: string
      form_settings:
        $ref: '#/definitions/dto.PublicFormSettings'
      id:
        type: string
      is_template:
        type: boolean
      questions:
        items:
          $ref: '#/definitions/dto.PublicQuestionDTO'
        type: array
      status:
        $ref: '#/definitions/domain.FormStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/domain.FormType'
    type: object
  dto.PublicFormSettings:
    properties:
      allow_backtrack:
        type: boolean
      auto_active_days:
        type: integer
      cover_image_url:
        type: string
      duration_minutes:
        type: integer
      end_time:
        type: string
      font_family:
        type: string
      fullscreen_mode:
        type: boolean
      is_active_immediately:
        type: boolean
      is_one_time_submission:
        type: boolean
      is_token_protected:
        type: boolean
      logo_url:
        type: string
      randomize_options:
        type: boolean
      randomize_questions:
        type: boolean
      show_question_number:
        type: boolean
      start_time:
        type: string
      theme_color:
        type: string
    type: object
  dto.PublicOptionDTO:
    properties:
      audio_url:
        type: string
      id:
        type: string
      img_url:
        type: string
      match_key:
        type: string
      match_target_text:
        type: string
      option_text:
        type: string
      order_index:
        type: integer
      video_url:
        type: string
    type: object
  dto.PublicQuestionDTO:
    properties:
      audio_url:
        type: string
      code_language:
        type: string
      id:
        type: string
      img_url:
        type: string
      is_auto_scored:
        type: boolean
      is_required:
        type: boolean
      options:
        items:
          $ref: '#/definitions/dto.PublicOptionDTO'
        type: array
      order_index:
        type: integer
      points:
        type: integer
      question_text:
        type: string
      question_type:
        $ref: '#/definitions/domain.QuestionType'
      video_url:
        type: string
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.ResendOTPRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  dto.SessionQuestionItemDTO:
    properties:
      answer_text:
        type: string
      is_answered:
        type: boolean
      is_flagged:
        type: boolean
      match_pairs:
        items:
          $ref: '#/definitions/dto.MatchPairItem'
        type: array
      order_index:
        type: integer
      question_id:
        type: string
      selected_option_id:
        type: string
    type: object
  dto.SessionStateDTO:
    properties:
      current_question_index:
        type: integer
      duration_minutes:
        type: integer
      form_id:
        type: string
      is_warning_acknowledged:
        type: boolean
      questions:
        items:
          $ref: '#/definitions/dto.SessionQuestionItemDTO'
        type: array
      response_id:
        type: string
      started_at:
        type: string
      status:
        type: string
      warning_message:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
    type: object
  dto.VerifyExamTokenRequest:
    properties:
      respondent_email:
        type: string
      token:
        type: string
    required:
    - respondent_email
    - token
    type: object
  dto.VerifyExamTokenResponse:
    properties:
      form:
        $ref: '#/definitions/dto.PublicFormDTO'
      response_id:
        type: string
      session_state:
        $ref: '#/definitions/dto.SessionStateDTO'
    type: object
  dto.VerifyOTPRequest:
    properties:
      email:
        type: string
      otp_code:
        type: string
    required:
    - email
    - otp_code
    type: object
  response.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@hidocs.id
    name: HiDocs Support
  description: High-performance backend API for HiDocs Form & Exam Maker application
    using Golang and Domain-Driven Design (DDD).
  termsOfService: http://swagger.io/terms/
  title: HiDocs Backend API (Form & Exam Maker)
  version: "1.0"
paths:
...
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Generate password reset token
      parameters:
      - description: Forgot Password Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Request password reset
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account. Generates a 6-digit OTP stored in
        Redis for 180 seconds.
      parameters:
      - description: Register Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Register user & send 6-digit OTP code to email
      tags:
      - Auth
  /api/v1/auth/resend-otp:
    post:
      consumes:
      - application/json
      parameters:
      - description: Resend OTP Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Resend 6-digit OTP code
      tags:
      - Auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using token
      parameters:
      - description: Reset Password Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Reset password
      tags:
      - Auth
  /api/v1/auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Validate 6-digit OTP code stored in Redis and return JWT Token
      parameters:
      - description: Verify OTP Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Verify OTP code and activate registration
      tags:
      - Auth
...
  /api/v1/public/forms/{form_id}/verify-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Verify Token Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyExamTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VerifyExamTokenResponse'
              type: object
      summary: Verify exam passcode token and start/resume exam session
      tags:
      - Public
  /api/v1/public/forms/{short_code}:
    get:
      parameters:
      - description: Custom URL Slug or Form ID
        in: path
        name: short_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PublicFormDTO'
              type: object
      summary: Access public form by custom URL slug or short code
      tags:
      - Public
  /api/v1/public/forms/{short_code}/qr:
    get:
      parameters:
      - description: Custom URL Slug or Form ID
        in: path
        name: short_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Generate QR Code URL for public form access
      tags:
      - Public
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"